Make collection utils common
