+++ src/utils/collectionUtils/listUtils.ts
@@ -1,3 +1,6 @@
+import { invariantFailed } from "../errors/throwErrors";
+import { stringifyAnyValue } from "../printers";
+
 import { EqualitySet } from "./equalitySet";
 import { EqualTo, HashUtils } from "./equalityUtils";
 
@@ -77,4 +80,15 @@ function makeAnyElementsUniqueImpl<T, K>(
         elements.map((element) => new WrappedT(element))
     );
     return set.elements().map((wrappedElement) => wrappedElement.element);
+}
+
+export function zip<T, V>(ts: T[], vs: V[]): [T, V][] {
+    if (ts.length !== vs.length) {
+        invariantFailed(
+            "Zip function",
+            "arrays should be of the same length, ",
+            `but got ${stringifyAnyValue(ts)} and ${stringifyAnyValue(vs)}`
+        );
+    }
+    return ts.map((t, i) => [t, vs[i]]);
 }
