+++ src/utils/collectionUtils/simpleSet.ts
